pyproject.toml
[tool.poetry]
name = "couting_stars"
version = "0.1.0"
description = "tarken tech challange"
authors = ["maiaatala <ana.atala@unemat.br>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.3.0"
opencv-python = "^4.6.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
icecream = "^2.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
